Added files, plus JsonServlet(?)
