+++ cookie-clicker-lab4-7/cookie-clicker-lab4-7/cookie-clicker-lab4-7/build.gradle
@@ -11,6 +11,7 @@ repositories {
 
 // In this section you declare the dependencies for your production and test code
 dependencies {
+compile 'com.google.code.gson:gson:2.8.2'
     // The production code uses the SLF4J logging API at compile time
     compile 'org.slf4j:slf4j-api:1.7.25'
     // Servlet
