+++ cookie-clicker-lab4-7/cookie-clicker-lab4-7/src/main/java/edu/csula/models/Generator.java
@@ -1,12 +1,12 @@
 package edu.csula.models;
 
 public class Generator {
-	private final int id;
-	private final String name;
-	private final String description;
-	private final int rate;
-	private final int baseCost;
-	private final int unlockAt;
+	private int id;
+	private String name;
+	private String description;
+	private int rate;
+	private int baseCost;
+	private int unlockAt;
 
 	public Generator(int id, String name, String description, int rate, int baseCost, int unlockAt) {
 		this.id = id;
@@ -20,24 +20,60 @@ public class Generator {
 	public int getId() {
 		return id;
 	}
+	public void setId(int id) {
+		this.id = id;
+	}
 
 	public String getName() {
 		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
 	}
 
 	public String getDescription() {
 		return description;
 	}
+	public void setDescription(String description) {
+		this.description = description;
+	}
 
 	public int getRate() {
 		return rate;
+	}
+	public void setRate(int rate) {
+		this.rate = rate;
 	}
 
 	public int getBaseCost() {
 		return baseCost;
 	}
+	public void setBaseCost(int baseCost) {
+		this.baseCost = baseCost;
+	}
 
 	public int getUnlockAt() {
 		return unlockAt;
+	}
+	public void setUnlockAt(int unlockAt) {
+		this.unlockAt = unlockAt;
+	}
+
+	@Override
+	public boolean equals(Object obj) {
+		if (obj == null) {
+			return false;
+		}
+		final Generator generator = (Generator) obj;
+		return this.id == generator.id && this.name.equals(generator.name) &&
+			this.description.equals(generator.description) &&
+			this.baseCost == generator.baseCost &&
+			this.rate == generator.rate &&
+			this.unlockAt == generator.unlockAt;
+	}
+
+	@Override
+	public int hashCode() {
+		return this.id;
 	}
 }
