+++ cookie-clicker-lab4-7/cookie-clicker-lab4-7/src/main/java/edu/csula/web/HelloJsonServlet.java
@@ -1,0 +1,29 @@
+package edu.csula.web;
+
+import java.io.IOException;
+import java.io.PrintWriter;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import edu.csula.models.User;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+
+@WebServlet("/hellojson")
+public class HelloJsonServlet extends HttpServlet {
+	public void doGet( HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		response.setContentType("application/json");
+		PrintWriter out = response.getWriter();
+
+		User u = new User(0, "username", "password");
+		GsonBuilder builder = new GsonBuilder();
+		Gson gson = builder.create();
+		String jsonString = gson.toJson(u);
+
+		out.println(jsonString);
+	}
+}
